fuck x

+++ src/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: ckurt <ckurt@student.42lyon.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/12 13:02:03 by ckurt             #+#    #+#             */
-/*   Updated: 2020/12/22 12:16:25 by ckurt            ###   ########lyon.fr   */
+/*   Updated: 2020/12/22 16:49:43 by ckurt            ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ int	select_parsing(va_list valist, t_element *elem)
 	if (elem->type == 'p')
 		return (ft_print_p(elem, va_arg(valist, size_t)));
 	if (elem->type == 'x')
-		return (ft_print_x(elem, va_arg(valist, size_t), "0123456789abcdef"));
+		return (ft_print_x(elem, va_arg(valist, unsigned long long), "0123456789abcdef"));
 	if (elem->type == 'X')
 		return (ft_print_x(elem, va_arg(valist, size_t), "0123456789ABCDEF"));
 	return (0);
